Created AudioFile class with basic functionality.
Added the wavetool.py file to implement the AudioFile class where the real work gets done. Also created wavetool_dev.py which can be used in place of the GUI for test and development. The updated code now has a real working feature: you can open a stereo wavefile and save it as two mono wavefiles.