This version has been V&V'd for compatibility with 32-bit floating point format wavefiles.
I generated a 32-bit floating point format wavefile of a mid-side guitar recording in Pro Tools. I V&V'd that the code can properly read the file, do the mid-side processing and write the results to output files in the correct format. I opened the results in Pro Tools and verified that the processing is correct (e.g., the two side channels cancel when panned together) and that the audio sounds consistent with a mid-side processed version generated entirely in Pro Tools.  I also tested the rumble filter in 32-bit float format and it seems to work fine in terms of the file I/O and file format, though I didn't conduct an audio test to ensure the filter is implemented.